[BAEL-1536] A Guide to the Resource Bundle #4
Essential changes after the second review:
- Doubles to BigDecimals

+++ core-java/src/main/java/com/baeldung/resourcebundle/ExampleResource_pl_PL.java
@@ -1,5 +1,6 @@
 package com.baeldung.resourcebundle;
 
+import java.math.BigDecimal;
 import java.util.ListResourceBundle;
 
 public class ExampleResource_pl_PL extends ListResourceBundle {
@@ -8,7 +9,7 @@ public class ExampleResource_pl_PL extends ListResourceBundle {
     protected Object[][] getContents() {
         return new Object[][] { 
             { "currency", "polish zloty" }, 
-            { "toUsdRate", new Double("3.401") }, 
+            { "toUsdRate", new BigDecimal("3.401") },
             { "cities", new String[] { "Warsaw", "Cracow" } } 
         };
     }
